Deprecate BaseRule
