+++ src/Attributes/Validate.php
@@ -3,10 +3,10 @@
 namespace Livewire\Attributes;
 
 use Attribute;
-use Livewire\Features\SupportValidation\BaseRule;
+use Livewire\Features\SupportValidation\BaseValidate;
 
 #[Attribute(Attribute::IS_REPEATABLE | Attribute::TARGET_ALL)]
-class Validate extends BaseRule
+class Validate extends BaseValidate
 {
     //
 }
